+++ src/bsocket.cc
@@ -122,7 +122,7 @@ bsocket &bsocket::operator=(bsocket &src)
 }
 
 int
-bsocket::baccept(unsigned long *host, int *port)
+bsocket::baccept(uint32_t *host, int *port)
 {
     return 0;
 }
@@ -313,20 +313,20 @@ bsocket::bselect(time_t outtime)
 int
 bsocket::bconnect(const char *host, int port)
 {
-    unsigned long rhost = inet_addr(host);
+    uint32_t rhost = inet_addr(host);
     if (inet_addr(host) == INADDR_NONE){
         hostent *phost = gethostbyname(host);
         if (phost == NULL){
             return -1;
         }
-        rhost = *(u_long*)phost->h_addr;
+        rhost = *(uint32_t *)phost->h_addr;
         fprintf(stderr, "IP: %s\n", inet_ntoa(*(in_addr*)&rhost));
     }
     return bconnect(rhost, port);
 }
 
 int
-bsocket::bconnect(unsigned long host, int port)
+bsocket::bconnect(uint32_t host, int port)
 {
     int error;
     int fflag;
@@ -361,7 +361,7 @@ bsocket::bconnect(unsigned long host, int port)
 
 int
 bsocket::bsendto(const void* buffer, size_t len,
-        unsigned long host, unsigned short port)
+        uint32_t host, unsigned short port)
 {
     int fflag;
     sockaddr_in siaddr;
@@ -379,7 +379,7 @@ bsocket::bsendto(const void* buffer, size_t len,
 
 int
 bsocket::brecvfrom(void* buffer, size_t len,
-        unsigned long *host, unsigned short *port)
+        uint32_t *host, unsigned short *port)
 {
     assert(b_fd != -1);
     sockaddr_in siaddr;
